pyproject.toml
[project]
name = "takatuski"
version = "0.1.0"
description = "請求書構造化システム"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.6",
    "openpyxl>=3.1.2",
    "pdfminer.six>=20221105",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.0.0",
    "openai>=1.12.0",
    "pytesseract>=0.3.10",
    "pillow>=11.1.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye.scripts]
dev = "uvicorn src.api.main:app --reload"

[tool.hatch.build.targets.wheel]
packages = ["src"]
